refactor: wallet -> pay package name
