+++ src/test/java/domain/wallet/WalletTest.java
@@ -2,6 +2,7 @@ package domain.wallet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.pay.Wallet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
